Feature: implemented downloading and managing executables from GitHub
