+++ internal/download/task.go
@@ -3,6 +3,7 @@ package download
 import (
 	"io"
 	"sync"
+	"time"
 
 	"github.com/wzhqwq/VRCDancePreloader/internal/utils"
 )
@@ -60,6 +61,8 @@ func newTaskWithoutManager(id string) *Task {
 		em: utils.NewEventManager[TaskChangeType](),
 
 		ID: id,
+
+		CancelCh: make(chan struct{}),
 	}
 }
 
@@ -104,9 +107,26 @@ func (t *Task) addBytes(size int64) {
 	t.em.NotifySubscribers(Progress)
 }
 
+func (t *Task) Speed() float64 {
+	if t.eta == nil {
+		return 0
+	}
+	return t.eta.QuerySpeed()
+}
+
+func (t *Task) RemainTime() time.Duration {
+	if t.eta == nil {
+		return -1
+	}
+	return t.eta.QueryRemainTime()
+}
+
 // Destroy
 
 func (t *Task) Cancel() {
+	if t.CancelCh == nil {
+		return
+	}
 	close(t.CancelCh)
 }
