+++ internal/third_party_api/settings.go
@@ -3,3 +3,4 @@ package third_party_api
 var YoutubeApiKey string
 var EnableYoutubeApi bool
 var EnableYoutubeThumbnail bool
+var EnableYtDlp bool
